Enable beta composer packages and experimental PHP for Travis

+++ .travis/before_install.php
@@ -3,5 +3,12 @@
 
 include_once 'common.php';
 
-// Disable XDebug
-run('phpenv config-rm xdebug.ini');
+// Disable XDebug for non-experimental PHP environments
+if (isNonExperimentalPhp()) {
+    run('phpenv config-rm xdebug.ini');
+}
+
+// Set composer minimum-stability configuration filter to beta versions
+if (usesBetaDependencies()) {
+    run('perl -pi -e \'s/^}$/,"minimum-stability":"beta"}/\' composer.json');
+}
